add magic numbers to serialization

+++ include/bdsg/hash_graph.hpp
@@ -312,7 +312,7 @@ public:
     ////////////////////////////////////////////////////////////////////////////
     
     /// Returns a static high-entropy number to indicate the class
-    uint64_t get_magic_number() const;
+    uint32_t get_magic_number() const;
     
 private:
